Move tt_memory.h out of api/tt-metalium
tt_memory.h represents the details of the memory representing a program binary,
so it doesn't need to be exposed to API consumers.

tt_memory is mainly used when interacting with a risc binary. A lot of that
code is in tt_metal/llrt/llrt.hpp, so move the header to tt_metal/llrt.
