+++ tt_metal/CMakeLists.txt
@@ -92,7 +92,6 @@ target_sources(
             api/tt-metalium/tile.hpp
             api/tt-metalium/tt_align.hpp
             api/tt-metalium/tt_backend_api_types.hpp
-            api/tt-metalium/tt_memory.h
             api/tt-metalium/tt_metal.hpp
             api/tt-metalium/util.hpp
             api/tt-metalium/utils.hpp
