+++ tt_metal/api/tt-metalium/kernel.hpp
@@ -28,7 +28,6 @@
 #include <tt-metalium/kernel_types.hpp>
 #include <tt-metalium/runtime_args_data.hpp>
 #include <tt-metalium/tt_backend_api_types.hpp>
-#include <tt-metalium/tt_memory.h>
 #include <umd/device/tt_core_coordinates.h>
 #include <umd/device/types/cluster_descriptor_types.h>
 #include <umd/device/types/xy_pair.h>
